% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_value_presence}
\alias{check_value_presence}
\title{Check for value presence}
\usage{
check_value_presence(val, vect, warn_msg = "")
}
\arguments{
\item{val}{element of vector \code{vect}}

\item{vect}{vector of values}

\item{warn_msg}{warning message to be displayed if \code{val} not in \code{vect}}
}
\value{
updated val
}
\description{
If value is not present in vector it takes its first value.
}
\keyword{internal}
