% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{column_to_row}
\alias{column_to_row}
\title{Column to row}
\usage{
column_to_row(data, colname)
}
\arguments{
\item{data}{data.frame with data}

\item{colname}{character with column name}
}
\value{
data.frame with one column less
}
\description{
Returns the same data.frame where one column is a rowname now.
}
