% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{combine_translations}
\alias{combine_translations}
\title{Combine two translation json files}
\usage{
combine_translations(x, y)
}
\arguments{
\item{x}{\code{chr/list} A path to a file or a list via \link[jsonlite]{fromJSON}}

\item{y}{\code{chr/list} A path to a file or a list via \link[jsonlite]{fromJSON}}
}
\value{
\code{list} combined translation object
}
\description{
Combine two translation json files
}
