% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{create_translation_file}
\alias{create_translation_file}
\title{Create translation file}
\usage{
create_translation_file(path, type = "json", handle = "i18n", output = NULL)
}
\arguments{
\item{path}{character with path of the file that needs to be inspected for
key translations}

\item{type}{type of the example output file with translations, either "json"
or "csv"}

\item{handle}{name of \code{Translator} object within script from \code{path}}

\item{output}{if NULL (default) the output will be saved with a default file
name ("translation.json" for JSON and "translate_lang.csv" for CSV)}
}
\description{
Auxiliary shiny.i18n function that searches for all key expressions (e.g.
surrounded by \code{i18n$t()} tag in the script).
}
