% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.i18_config}
\alias{.i18_config}
\title{Global  i18n config list.}
\format{
An object of class \code{list} of length 3.
}
\usage{
.i18_config
}
\description{
Global  i18n config list.
}
\keyword{datasets}
