% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\docType{data}
\name{.translator_options}
\alias{.translator_options}
\title{Translator options}
\format{
An object of class \code{list} of length 3.
}
\usage{
.translator_options
}
\description{
Translator options
}
\keyword{internal}
