% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{extract_key_expressions}
\alias{extract_key_expressions}
\title{Extract key expressions}
\usage{
extract_key_expressions(text, handle = "i18n")
}
\arguments{
\item{text}{character with text of the script}

\item{handle}{character with Translator object handle (default: 'i18n')}
}
\value{
vector of characters with key expressions
}
\description{
Extract key expressions
}
