% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_i18n_config}
\alias{get_i18n_config}
\title{Get i18n config}
\usage{
get_i18n_config(field)
}
\arguments{
\item{field}{a field from configuration file}
}
\value{
character with option from config.yaml file
}
\description{
Get i18n config
}
\keyword{internal}
