% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{i18n_state}
\alias{i18n_state}
\title{Create i18n state div}
\usage{
i18n_state(init_language)
}
\arguments{
\item{init_language}{key language code}
}
\value{
shiny tag with div \code{"i18n-state"}
}
\description{
This hidden div contain information about i18 translation state.
}
