% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{init_i18n}
\alias{init_i18n}
\title{Creates new i18n Translator object}
\usage{
init_i18n(
  translation_csvs_path = NULL,
  translation_json_path = NULL,
  translation_csv_config = NULL,
  automatic = FALSE
)
}
\arguments{
\item{translation_csvs_path}{character with path to folder containing csv
translation files. See more in  Details.}

\item{translation_json_path}{character with path to JSON translation file.
See more in  Details.}

\item{translation_csv_config}{character with path to configuration file for
csv option.}

\item{automatic}{logical flag, indicating if i18n should use an automatic
translation API.}
}
\value{
i18n object
}
\description{
Creates new i18n Translator object
}
\examples{
\dontrun{
 i18n <- init_i18n(translation_csvs_path = "../csvdata/")
 i18n <- init_i18n(translation_json_path = "translations.json")
}
}
