% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_local_config}
\alias{load_local_config}
\title{Load Local YAML Config}
\usage{
load_local_config(yaml_config_path)
}
\arguments{
\item{yaml_config_path}{path to yaml config file}
}
\value{
list of config options or empty list if file not exists
}
\description{
Load Local YAML Config
}
