% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multmerge}
\alias{multmerge}
\title{Multiple Merge}
\usage{
multmerge(filenames, sep = ",")
}
\arguments{
\item{filenames}{character vector with filenames}

\item{sep}{fields separator}
}
\value{
data.frame with merged files
}
\description{
Inspired by:
https://www.r-bloggers.com/merging-multiple-data-files-into-one-data-frame/
}
\keyword{internal}
