% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_and_merge_csvs}
\alias{read_and_merge_csvs}
\title{Read and merge CSVs}
\usage{
read_and_merge_csvs(dir_path)
}
\arguments{
\item{dir_path}{character with path to directory with csv files}
}
\value{
data.frame with CSV files content
}
\description{
This function reads and merges data from multiple csv files in given folder.
}
