% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_translation.R
\name{read_translation}
\alias{read_translation}
\title{Read translation from file}
\usage{
read_translation(file, fallback_lang = "en")
}
\arguments{
\item{path}{String, localization of file with translations.}
}
\value{
List with key-value pairs of attributes in desired language.
}
\description{
\code{read_translation} returns list with key-value pairs of attributes in
desired language.
}
\details{
This function reads predefined file which contains attributes (keys) and
translations (values). At the moment function can read only CSV files. JSON
and perhaps other formats will be implemented soon.
}
\examples{
en <- read_translation("data/translations.csv", "en")
en$page_title
pl <- read_translation("data/translations.csv", "pl")
pl$page_title

\dontrun{
read_translation("data/translation.csv")
}

}
