% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{save_to_csv}
\alias{save_to_csv}
\title{Save example i18n file to CSV}
\usage{
save_to_csv(key_expressions, output_path = NULL)
}
\arguments{
\item{key_expressions}{vector with key expression to translate}

\item{output_path}{character with path to output file (default:
"translate_lang.csv" if NULL)}
}
\description{
It saves translation data structure with language key code "key" and
language to translate name "lang" as an example of a translation csv file.
}
