% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{save_to_json}
\alias{save_to_json}
\title{Save example i18n file to json}
\usage{
save_to_json(key_expressions, output_path = NULL)
}
\arguments{
\item{key_expressions}{vector with key expression to translate}

\item{output_path}{character with path to output file (default:
"translation.json" if NULL)}
}
\description{
It saves translation data structure with language key code "key" as an
example of a translation JSON file.
}
