% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.i18n-package.R
\docType{package}
\name{shiny.i18n-package}
\alias{shiny.i18n}
\alias{shiny.i18n-package}
\title{shiny.i18n: Shiny Applications Internationalization}
\description{
It provides easy internationalization of Shiny applications. It can be used as standalone translation package to translate reports, interactive visualizations or graphical elements as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://appsilon.github.io/shiny.i18n/}
  \item \url{https://github.com/Appsilon/shiny.i18n}
  \item Report bugs at \url{https://github.com/Appsilon/shiny.i18n/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowicki \email{opensource+kuba@appsilon.com}

Authors:
\itemize{
  \item Dominik Krzemiński \email{raymon92@gmail.com}
  \item Krystian Igras \email{krystian8207@gmail.com}
  \item Jakub Sobolewski \email{jakub.sobolewski@appsilon.com}
}

Other contributors:
\itemize{
  \item Appsilon Sp. z o.o. \email{opensource@appsilon.com} [copyright holder]
}

}
\keyword{internal}
