% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.i18n.R
\docType{package}
\name{shiny.i18n}
\alias{shiny.i18n}
\alias{shiny.i18n-package}
\title{shiny.i18n: A package for Shiny applications internationalization.}
\description{
The shiny.i18n package provides functions for easy Shiny applications internationalization.#'
}
\section{Foo functions}{

The foo functions ...
}

