% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic.R
\name{translate_with_google_cloud}
\alias{translate_with_google_cloud}
\title{This provides functions for automatic translations with online APIs
Translate with Google cloud}
\usage{
translate_with_google_cloud(txt_to_translate, target_lang)
}
\arguments{
\item{txt_to_translate}{character with text to translate}

\item{target_lang}{character with language code}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is wrapper for \code{gl_translate} function from \code{googleLanguageR}
package.
}
