% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{Translator}
\alias{Translator}
\title{Translator R6 Class}
\description{
Translator R6 Class

Translator R6 Class
}
\details{
This creates shinny.i18n Translator object used for translations.
Now you can surround the pieces of the text you want to translate by
one of the translate statements (ex.: \code{Translator$t("translate me")}).
Find details in method descriptions below.
}
\examples{
\dontrun{
  i18n <- Translator$new(translation_json_path = "translation.json") # translation file
  i18n$set_translation_language("it")
  i18n$t("This text will be translated to Italian")
}

# Shiny example
if (interactive()) {
library(shiny)
library(shiny.i18n)
 #to run this example  make sure that you have a translation file
 #in the same path
i18n <- Translator$new(translation_json_path = "examples/data/translation.json")
i18n$set_translation_language("pl")
ui <- fluidPage(
  h2(i18n$t("Hello Shiny!"))
)
server <- function(input, output) {}
shinyApp(ui = ui, server = server)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Translator$new()}}
\item \href{#method-get_dict}{\code{Translator$get_dict()}}
\item \href{#method-get_languages}{\code{Translator$get_languages()}}
\item \href{#method-get_language}{\code{Translator$get_language()}}
\item \href{#method-translate}{\code{Translator$translate()}}
\item \href{#method-t}{\code{Translator$t()}}
\item \href{#method-set_language}{\code{Translator$set_language()}}
\item \href{#method-clone}{\code{Translator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Translator with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$new(translation_json_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{translation_json_path}}{character with path to JSON translation file.
See more in  Details.}

\item{\code{translation_csvs_path}}{character with path to folder containing csv
translation files. Files must have "translation_" prefix, for example:
\code{translation_<LANG-CODE>.csv}.}

\item{\code{translation_csv_config}}{character with path to configuration file for
csv option.}

\item{\code{separator_csv}}{separator of CSV values (default ",")}

\item{\code{automatic}}{logical flag, indicating if i18n should use an automatic
translation API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dict"></a>}}
\if{latex}{\out{\hypertarget{method-get_dict}{}}}
\subsection{Method \code{get_dict()}}{
Get dictionnary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$get_dict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_languages"></a>}}
\if{latex}{\out{\hypertarget{method-get_languages}{}}}
\subsection{Method \code{get_languages()}}{
Get all available languages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$get_languages()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_language"></a>}}
\if{latex}{\out{\hypertarget{method-get_language}{}}}
\subsection{Method \code{get_language()}}{
Get current target translation language
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$get_language()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\if{latex}{\out{\hypertarget{method-translate}{}}}
\subsection{Method \code{translate()}}{
Translates 'keyword' to language specified by 'set_translation_language'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$translate(keyword, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keyword}}{character or vector of characters with a word or
expression to translate}

\item{\code{session}}{Shiny server session (default: current reactive domain)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-t"></a>}}
\if{latex}{\out{\hypertarget{method-t}{}}}
\subsection{Method \code{t()}}{
Wrapper to \code{translate} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$t(keyword, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keyword}}{Character or vector of characters with a word or
expression to translate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_language"></a>}}
\if{latex}{\out{\hypertarget{method-set_language}{}}}
\subsection{Method \code{set_language()}}{
Specify language of translation. It must exist in 'languages' field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$set_language(language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{language}}{character with a translation language code}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
