% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_names}
\alias{validate_names}
\title{Validate Column Names}
\usage{
validate_names(list_df, n = 1)
}
\arguments{
\item{list_df}{list of data frames}

\item{n}{integer denoting column number}
}
\value{
TRUE if names of n-th columns of data.frames is the same,
FALSE otherwise.
}
\description{
Validate if n-th column name of data.frames (given in list) is the same.
}
\keyword{internal}
